Statistics: move histogram into Common
